backend/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14.0
django-filter>=23.0
django-crontab==0.7.1
mysqlclient>=2.2.0
Pillow
django-cors-headers>=4.0.0
python-dotenv>=1.0.0
drf-yasg>=1.21.7

# --- Tests ---
pytest>=7.4.0
pytest-django>=4.5.0
factory-boy>=3.3.0